requirements.txt
pyyaml==6.0.1
sympy==1.12
pymodbus==3.6.3